// src/LawyerSite.Web/src/app/services/client/web-api-client.ts
export interface CreateBlockedSlotDto {
    dateTime?: Date;
    durationHours?: number;
    durationMinutes?: number;
    reason?: string | undefined;
    comment?: string | undefined;
    isFullDay?: boolean;
}